\name{show_tayloR_palette}
\alias{show_tayloR_palette}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
show_tayloR_palette
}
\description{
Displays a tayloR_palette in plots with name of palette
}
\usage{
show_tayloR_palette(palette_name)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{palette_name}{
palette_name is the name of the palette you'd like to see (e.g., "debut")
}
}
\author{
Bridget Chalifour
}
