\name{tayloR_palette}
\alias{tayloR_palette}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
tayloR_palette
}
\description{
This function create color palettes.
}
\usage{
tayloR_palette(palette_name)
}
tayloR_palette(palette_name = palette name)
\arguments{
  \item{palette_name}{
Name of palette you would like to use
}
}

\author{
Bridget Chalifour
}
